test(types): move tests combining multiple operators to integration testing
