+++ compiler/fluxc_types/Cargo.toml
@@ -2,3 +2,6 @@
 name = "fluxc_types"
 version = "0.1.0"
 edition = "2021"
+
+[dev-dependencies]
+pretty_assertions = "1"
