+++ compiler/fluxc_types/src/operator/union.rs
@@ -29,6 +29,12 @@ impl Unify<Type> for Type {
     }
 }
 
+impl Into<Type> for Union {
+    fn into(self) -> Type {
+        self.simplify()
+    }
+}
+
 impl Simplify for Union {
     fn simplify(&self) -> Type {
         let lhs = self.lhs.simplify();
@@ -51,5 +57,79 @@ impl Simplify for Union {
         } else {
             Type::Operation(Operation::Union(Union::of(lhs.into(), rhs.into())))
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use pretty_assertions::assert_eq;
+
+    use crate::{Operation, Primitive, Type, Unify, Union};
+
+    #[test]
+    fn unify_primitives() {
+        // string | int = string | int
+        assert_eq!(
+            Type::Primitive(Primitive::String).unify(&Type::Primitive(Primitive::Int)),
+            Type::Operation(Operation::Union(Union::of(
+                Type::Primitive(Primitive::String),
+                Type::Primitive(Primitive::Int),
+            )))
+        );
+        // string | string = string
+        assert_eq!(
+            Type::Primitive(Primitive::String).unify(&Type::Primitive(Primitive::String)),
+            Type::Primitive(Primitive::String)
+        );
+    }
+
+    #[test]
+    fn unify_unions() {
+        // (string | int) | (string | int) = (string | int)
+        assert_eq!(
+            Type::Operation(Operation::Union(Union::of(
+                Type::Operation(Operation::Union(Union::of(
+                    Type::Primitive(Primitive::String),
+                    Type::Primitive(Primitive::Int),
+                ))),
+                Type::Operation(Operation::Union(Union::of(
+                    Type::Primitive(Primitive::String),
+                    Type::Primitive(Primitive::Int),
+                ))),
+            )))
+            .unify(&Type::Operation(Operation::Union(Union::of(
+                Type::Primitive(Primitive::String),
+                Type::Primitive(Primitive::Int),
+            )))),
+            Type::Operation(Operation::Union(Union::of(
+                Type::Primitive(Primitive::String),
+                Type::Primitive(Primitive::Int),
+            )))
+        );
+        // (string | int) | (string | float) = (string | int | float)
+        assert_eq!(
+            Type::Operation(Operation::Union(Union::of(
+                Type::Operation(Operation::Union(Union::of(
+                    Type::Primitive(Primitive::String),
+                    Type::Primitive(Primitive::Int),
+                ))),
+                Type::Operation(Operation::Union(Union::of(
+                    Type::Primitive(Primitive::String),
+                    Type::Primitive(Primitive::Float),
+                ))),
+            )))
+            .unify(&Type::Operation(Operation::Union(Union::of(
+                Type::Primitive(Primitive::String),
+                Type::Primitive(Primitive::Int),
+            )))),
+            Type::Operation(Operation::Union(Union::of(
+                Union::of(
+                    Type::Primitive(Primitive::String),
+                    Type::Primitive(Primitive::Int),
+                )
+                .into(),
+                Type::Primitive(Primitive::Float),
+            )))
+        );
     }
 }
